+++ packages/components/theme/Login/index.ts
@@ -1,8 +1,0 @@
-import Login from "./src/index.vue";
-
-export { Login as TkLogin };
-export default Login;
-
-export * from "./src/login";
-export * from "./src/useWatchLogin";
-export * from "./src/instance";