+++ packages/components/theme/Login/src/instance.ts
@@ -1,3 +1,0 @@
-import type Login from "./index.vue";
-
-export type TkLoginInstance = InstanceType<typeof Login>;