+++ packages/locale/lang/en.ts
@@ -104,7 +104,7 @@ export default {
       infoLabel: "Article Information",
       coverImgLabel: "Article Cover Image",
     },
-    homeCard: {
+    pageCard: {
       label: "Home Card",
       prev: "Previous Page",
       next: "Next Page",
@@ -269,6 +269,15 @@ export default {
       loginInfoNull: "Please input your username and password!",
       usernamePlaceholder: "Please input a username",
       passwordPlaceholder: "Please input a password",
+      verifyCodePlaceholder: "Please input the verification code",
+      verifyCodeNonNull: "Please input the verification code",
+      verifyCodeError: "Please input the correct verification code",
+    },
+    riskLink: {
+      label: "Risk Link Page",
+      title: "Leaving {name} soon, please pay attention to property safety",
+      linkIllegal: "Link security verification in progress, please wait ...",
+      confirmButtonText: "Continue to visit",
     },
   },
 };